// src/api/types.gen.ts
export type BenchmarkMode = 'rating' | 'static' | 'live';
// ...
export type LlmContestProblemSubmission = {
    problem_index: string;
    problem_link: string;
    problem_name: string;
    problem_extra: string;
    verdict: 'accepted' | 'not_accepted' | 'unknown' | 'in_queue' | 'generation_failed' | 'in_batch' | 'compilation_error';
    status: string | null;
    submission_id: string | null;
};

export type LlmContestSubmissions = {
    contest_title: string;
    contest_link: string;
    contest_start_time: number;
    problems: Array<LlmContestProblemSubmission>;
};

export type LlmDifficulty = {
    llms: Array<LlmByDifficulty>;
    difficulty: ProblemDifficulty;
};

export type LlmSubmissions = {
    difficulty?: ProblemDifficulty;
    contests: Array<LlmContestSubmissions>;
};

export type ProblemDifficulty = 'easy' | 'medium' | 'hard' | 'all';